1. Renamed base_GFormula.py to base.py 2.Renamed class name GFormula to BaseGMethod and moved it to base.py 3. Added treatment_strategy in the estimate_individual_outcome and estimate_population_outcome methods. 4. moved _apply_noise from abstract class to subclass
